big mess.  Think i have to make a profile table.
Having a user for devise and then trying to have an existing create
their user once signed up is causing all sorts of errors because I
can’t call User.new without losing the current info with my user.  So
have to figure out if we just want to prompt the user for all the info
off the jump or just see if I can create a profile table.